AI_Property_Investement_Agent/requirements.txt
langchain 
langchain_community 
crewai 
crewai_tools==0.4.8
streamlit 
python-dotenv